Rename AddressBook classes and methods to CLientHub
